requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
openai==1.60.0
python-dotenv==1.0.1
datasets==3.0.0
websockets==13.0
httpx==0.27.0
